AnnData to/from SingleCellExperiment

Converts a Python AnnData object to or from a SingleCellExperiment object.

AnnData2SCE(adata)

SCE2AnnData(sce, X_name = NULL)

Arguments

adata
A reticulate reference to a Python AnnData object.

sce
A SingleCellExperiment object.

X_name
Name of the assay to use as the primary matrix (X) of the AnnData object. If NULL, the first assay of sce will be used by default.

Value

AnnData2SCE will return a SingleCellExperiment containing the equivalent data from adata.

SCE2AnnData will return a reticulate reference to an AnnData object containing the content of sce.

Details

These functions assume that an appropriate Python environment has already been loaded. As such, they are largely intended for developer use, most typically inside a basilisk context.

The conversion is not entirely lossless. No attempt is made by AnnData2SCE to transfer the alternative Experiments from sce to an AnnData object. Conversely, values in the obsm field of adata are not transferred to a SingleCellExperiment.

See also

Author

Luke Zappia

Examples

library(basilisk)
library(scRNAseq)

#> Loading required package: SingleCellExperiment

#> Loading required package: SummarizedExperiment

#> Loading required package: GenomicRanges

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which.max, which.min

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> Loading required package: GenomeInfoDb

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: DelayedArray

#> Loading required package: Matrix

#> 
#> Attaching package: ‘Matrix’

#> The following object is masked from ‘package:S4Vectors’:
#> 
#>     expand

#> Loading required package: matrixStats

#> 
#> Attaching package: ‘matrixStats’

#> The following objects are masked from ‘package:Biobase’:
#> 
#>     anyMissing, rowMedians

#> 
#> Attaching package: ‘DelayedArray’

#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colMaxs, colMins, colRanges, rowMaxs, rowMins, rowRanges

#> The following objects are masked from ‘package:base’:
#> 
#>     aperm, apply, rowsum

seger <- SegerstolpePancreasData()

#> using temporary cache /tmp/Rtmpe2xL8v/BiocFileCache

#> snapshotDate(): 2020-06-11

#> see ?scRNAseq and browseVignettes('scRNAseq') for documentation

#> downloading 1 resources

#> retrieving 1 resource

#> loading from cache

#> see ?scRNAseq and browseVignettes('scRNAseq') for documentation

#> downloading 1 resources

#> retrieving 1 resource

#> loading from cache

#> see ?scRNAseq and browseVignettes('scRNAseq') for documentation

#> downloading 1 resources

#> retrieving 1 resource

#> loading from cache

# If you don't know what the code below is doing,
# you probably shouldn't be using these functions.
roundtrip <- basiliskRun(env=zellkonverter:::anndata_env, fun=function(sce) {
     # Convert SCE to AnnData:
     ad <- SCE2AnnData(sce) 

     # Maybe do some work in Python on 'ad':
     # BLAH BLAH BLAH

     # Convert back to an SCE:
     AnnData2SCE(ad)
}, sce=seger)

#> using the 'counts' assay as the X matrix

#> Warning: index contains duplicated values: row names not set