Write H5AD

Write a H5AD file from a SingleCellExperiment object.

writeH5AD(sce, file, X_name = NULL, skip_assays = FALSE)

Arguments

sce
A SingleCellExperiment object.

file
String containing a path to write the new .h5ad file.

X_name
Name of the assay to use as the primary matrix (X) of the AnnData object. If NULL, the first assay of sce will be used by default.

skip_assays
Logical scalar indicating whether assay matrices should be ignored when writing to file.

Value

A NULL is invisibly returned.

Details

Setting skip_assays=TRUE can occasionally be useful if the matrices in sce are stored in a format that is not amenable for efficient conversion to a numpy-compatible format. In such cases, it can be better to create an empty placeholder dataset in file and fill it in R afterwards.

When first run, this function will instantiate a conda environment containing all of the necessary dependencies. This will not be performed on any subsequent run or if any other zellkonverter function has been run prior to this one.

The anndata package automatically converts some character vectors to factors when saving .h5ad files. This can effect columns of rowData(sce) and colData(sce) which may change type when the .h5ad file is read back into R.

See also

Author

Luke Zappia

Aaron Lun

Examples

# Using the Zeisel brain dataset
if (requireNamespace("scRNAseq", quietly = TRUE)) {
    library(scRNAseq)
    sce <- ZeiselBrainData()

    # Writing to a H5AD file
    temp <- tempfile(fileext = ".h5ad")
    writeH5AD(sce, temp)
}

#> snapshotDate(): 2020-12-09

#> see ?scRNAseq and browseVignettes('scRNAseq') for documentation

#> downloading 1 resources

#> retrieving 1 resource

#> loading from cache

#> see ?scRNAseq and browseVignettes('scRNAseq') for documentation

#> downloading 1 resources

#> retrieving 1 resource

#> loading from cache

#> see ?scRNAseq and browseVignettes('scRNAseq') for documentation

#> downloading 1 resources

#> retrieving 1 resource

#> loading from cache

#> snapshotDate(): 2020-12-09

#> see ?scRNAseq and browseVignettes('scRNAseq') for documentation

#> loading from cache

#> Note: using the 'counts' assay as the X matrix